requirements.txt
appdirs==1.4.4
argon2-cffi==20.1.0
asgiref==3.4.1
astroid==2.6.2
async-generator==1.10
attrs==21.2.0
autopep8==1.5.7
backcall==0.2.0
bandit==1.7.0
bleach==3.3.0
boto3==1.18.40
botocore==1.21.40
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
colorama==0.4.4
cryptography==3.4.7
debugpy==1.3.0
decorator==5.0.9
defusedxml==0.7.1
distlib==0.3.2
dj-database-url==0.5.0
Django==3.2.7
django-allauth==0.41.0
django-ckeditor==6.2.0
django-countries==7.2.1
django-crispy-forms==1.12.0
django-js-asset==1.2.2
django-storages==1.11.1
docutils==0.17.1
entrypoints==0.3
filelock==3.0.12
flake8==3.9.2
flake8-django==1.1.2
flake8-flask==0.9.3
gitdb==4.0.7
GitPython==3.1.18
gunicorn==20.1.0
idna==2.10
importlib-metadata==4.6.1
ipykernel==6.0.1
ipython==7.25.0
ipython-genutils==0.2.0
isort==5.9.1
jedi==0.18.0
jeepney==0.6.0
Jinja2==3.0.1
jmespath==0.10.0
jsonschema==3.2.0
jupyter-client==6.1.12
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
keyring==23.0.1
lazy-object-proxy==1.6.0
MarkupSafe==2.0.1
matplotlib-inline==0.1.2
mccabe==0.6.1
mistune==0.8.4
mypy==0.910
mypy-extensions==0.4.3
nbclient==0.5.3
nbconvert==6.1.0
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.4.0
oauthlib==3.1.1
packaging==21.0
pandocfilters==1.4.3
parso==0.8.2
pbr==5.6.0
pep8==1.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.3.2
pipenv==2021.5.29
pkginfo==1.7.0
prometheus-client==0.11.0
prompt-toolkit==3.0.19
psycopg2==2.9.1
ptyprocess==0.7.0
pycodestyle==2.7.0
pycparser==2.20
pydocstyle==6.1.1
pyflakes==2.3.1
Pygments==2.9.0
PyJWT==2.1.0
pylama==7.7.1
pylint==2.9.3
pyparsing==2.4.7
pyrsistent==0.18.0
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2021.1
PyYAML==5.4.1
pyzmq==22.1.0
r2c-py-ast==0.1.0b1
readme-renderer==29.0
requests==2.25.1
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
rfc3986==1.5.0
rope==0.19.0
s3transfer==0.5.0
SecretStorage==3.3.1
Send2Trash==1.7.1
six==1.16.0
smmap==4.0.0
snowballstemmer==2.1.0
sqlparse==0.4.2
stevedore==3.3.0
stripe==2.60.0
terminado==0.10.1
testpath==0.5.0
toml==0.10.2
tornado==6.1
tqdm==4.61.2
traitlets==5.0.5
twine==3.4.1
typing-extensions==3.10.0.0
urllib3==1.26.6
virtualenv==20.4.7
virtualenv-clone==0.5.4
wcwidth==0.2.5
webencodings==0.5.1
wrapt==1.12.1
zipp==3.5.0
